refactor: deletado pasta que veio errado do merge + ajustado includes
