+++ user-service/src/Application/DTO/ToCommand/DeleteUserDTO.ts
