+++ user-service/src/Domain/Queries/QueryObject/GetUserQuery.ts
@@ -1,4 +1,4 @@
-import {GetUserDTO} from "../../../Application/DTO/Query/GetUserDTO.js";
+import {GetUserDTO} from "../../../Application/DTO/ToQuery/GetUserDTO.js";
 
 export class GetUserQuery
 {
