+++ user-service/src/Domain/Queries/QueryObject/VerifyIfUsersExistsByUuidsQuery.ts
@@ -1,4 +1,4 @@
-import {VerifyIfUsersExistsByUuidsDTO} from "../../../Application/DTO/Query/VerifyIfUsersExistsByUuidsDTO.js";
+import {VerifyIfUsersExistsByUuidsDTO} from "../../../Application/DTO/ToQuery/VerifyIfUsersExistsByUuidsDTO.js";
 
 export class VerifyIfUsersExistsByUuidsQuery
 {
