+++ user-service/src/Infrastructure/Persistence/Repositories/Interface/IBaseRepository.ts
@@ -1,7 +1,4 @@
 import {NotificationError} from "../../../../Shared/Errors/NotificationError.js";
-import {GetUserDTO} from "../../../../Application/DTO/Query/GetUserDTO.js";
-import {GetUserQueryDTO} from "../../../../Domain/QueryDTO/GetUserQueryDTO.js";
-import {User} from "../../../../Domain/Entities/Concrete/User.js";
 
 export interface IBaseRepository<TQueryDTO, TEntity>
 {
