+++ user-service/src/Presentation/Controllers/UserController.ts
@@ -1,22 +1,22 @@
-import { CreateUserDTO } from "../../Application/DTO/Command/CreateUserDTO.js";
+import { CreateUserDTO } from "../../Application/DTO/ToCommand/CreateUserDTO.js";
 import { Result } from "../../Shared/Utils/Result.js";
 import { FastifyReply, FastifyRequest } from "fastify";
 import { BaseController } from "./BaseController.js";
 import { CreateUserService } from "../../Application/Services/Concrete/CreateUserService.js";
 import { NotificationError } from "../../Shared/Errors/NotificationError.js";
-import { EditUserDTO } from "../../Application/DTO/Command/EditUserDTO.js";
+import { EditUserDTO } from "../../Application/DTO/ToCommand/EditUserDTO.js";
 import { EditUserService } from "../../Application/Services/Concrete/EditUserService.js";
-import { DeleteUserDTO } from "../../Application/DTO/Command/DeleteUserDTO.js";
+import { DeleteUserDTO } from "../../Application/DTO/ToCommand/DeleteUserDTO.js";
 import { DeleteUserService } from "../../Application/Services/Concrete/DeleteUserService.js";
-import { GetUserDTO } from "../../Application/DTO/Query/GetUserDTO.js";
 import { GetUserService } from "../../Application/Services/Concrete/GetUserService.js";
 import { GetUserViewModel } from "../../Application/ViewModels/GetUserViewModel.js";
-import {VerifyIfUsersExistsByUuidsDTO} from "../../Application/DTO/Query/VerifyIfUsersExistsByUuidsDTO.js";
 import {VerifyIfUserExistsByUsernameDTO} from "../../Application/DTO/ToQuery/VerifyIfUserExistsByUsernameDTO.js";
 import {VerifyIfUsersExistsByUsernamesDTO} from "../../Application/DTO/ToQuery/VerifyIfUsersExistsByUsernamesDTO.js";
 import {UpdateStatsDTO} from "../../Application/DTO/ToCommand/UpdateStatsDTO.js";
 import {UserRepository} from "../../Infrastructure/Persistence/Repositories/Concrete/UserRepository.js";
 import {UserService} from "../../Application/Services/Concrete/UserService.js";
+import {GetUserDTO} from "../../Application/DTO/ToQuery/GetUserDTO.js";
+import {VerifyIfUsersExistsByUuidsDTO} from "../../Application/DTO/ToQuery/VerifyIfUsersExistsByUuidsDTO.js";
 
 export class UserController extends BaseController {
   private readonly notificationError: NotificationError;
