+++ user-service/src/Presentation/Controllers/UserSessionController.ts
@@ -1,7 +1,7 @@
 import {BaseController} from "./BaseController.js";
 import {NotificationError} from "../../Shared/Errors/NotificationError.js";
 import {FastifyReply, FastifyRequest} from "fastify";
-import {UserSessionDTO} from "../../Application/DTO/Command/UserSessionDTO.js";
+import {UserSessionDTO} from "../../Application/DTO/ToCommand/UserSessionDTO.js";
 import {Result} from "../../Shared/Utils/Result.js";
 import {LoginUserService} from "../../Application/Services/Concrete/LoginUserService.js";
 import {LogoutUserService} from "../../Application/Services/Concrete/LogoutUserService.js";
